// sourcecode/BTrak.Packages/snova-testrepo/testrepo-module-components/testrepo-components/src/lib/snovasys-testrepo/models/testcaserundetails.ts
export class TestCaseRunDetails {
    testSuiteId: string;
    testCaseId: string;
    sectionId: string;
    casesSelected: number;
    casesCount: number;
    sectionSelected: boolean;
    sectionCheckBoxClicked: boolean;
    unselectSection: boolean;
    selectSection: boolean;
    multiSections: boolean;
    unselectAllCases: boolean;
    sectionsAllNone: boolean;
}